+++ app/appwrite/authentication/index.ts
@@ -1,0 +1,6 @@
+export { default as loginWithGoogle } from "./loginWithGoogle";
+export { default as getUser } from "./getUser";
+export { default as getGooglePicture } from "./getGooglePicture";
+export { default as logoutUser } from "./logoutUser";
+export { default as storeUserData } from "./storeUserData";
+export { default as getExistingUser } from "./getExistingUser";