+++ app/appwrite/authentication/logoutUser.ts
@@ -1,0 +1,11 @@
+import { account } from "~/appwrite/client";
+
+export const logoutUser = async () => {
+    try {
+        await account.deleteSession("current");
+    } catch(e) {
+        console.error("logoutUser:", e);
+    }
+}
+
+export default logoutUser;